sprite support works, but simple_effects.cpp are screwed up :> (imgfmts.cpp did not have bitmap flipping, tho it should have, and simple_effects.cpp assumed that the "old" behaviour of imgfmts.cpp was correct -- which sux)
git-svn-id: svn://ivan.vucica.net/theoutcast@12 4617c47b-1226-0410-b705-8e5dc674c1d5

+++ theoutcast/sprfmts.h
@@ -6,7 +6,7 @@
 bool SPRLoader(std::string sprfile);
 bool SPRUnloader();
 
-extern unsigned short *SPRPointers;
+extern unsigned long *SPRPointers;
 extern unsigned short SPRCount;
 extern std::string SPRFile;
